moved and split
