+++ src/server/service.js
@@ -2,6 +2,8 @@
 /* eslint-disable no-plusplus */
 const { promisify } = require('util');
 const { Transform } = require('stream');
+const fs = require('fs');
+const path = require('path');
 
 function myMap(callback) {
   const result = [];
@@ -108,9 +110,89 @@ function createCsvToJson() {
   return new Transform({ transform, flush });
 }
 
+function stringToObject(str, optimized) {
+  let strNew;
+  if (str.charAt(str.length - 1) !== ',') strNew = str.slice(3, -1);
+  else strNew = str.slice(3, str.length - 2);
+
+  let value = strNew.split(',');
+  value = value.map((strI) => strI.slice(strI.indexOf(':') + 2));
+  value[0] = value[0].slice(1, -1);
+  value[1] = value[1].slice(1, -1);
+  value[2] = Number(value[2]);
+  value[3] = Number(value[3]);
+
+  const product = { type: value[0], color: value[1], quantity: value[2], price: value[3] };
+  const index = `${product.type}_${product.color}_${product.price}`;
+
+  if (!optimized.has(index)) optimized.set(index, product);
+  else optimized.get(index).quantity += product.quantity;
+}
+
+function arrayObjectsToString(optimized) {
+  let result = [];
+  let totalQuantity = 0;
+  // eslint-disable-next-line no-restricted-syntax
+  for (const [, product] of optimized) {
+    result.push(
+      `  {"type": "${product.type}", "color": "${product.color}", ` +
+        `"quantity": ${product.quantity}, "price": ${product.price}}`,
+    );
+    totalQuantity += product.quantity;
+  }
+  result = `[\n${result.join(',\n')}\n]`;
+  return { result, totalQuantity };
+}
+
+function csvOptimization(fileName) {
+  return new Promise((resolve, reject) => {
+    let filePath = path.resolve('./upload/', fileName);
+
+    const streamReading = fs.createReadStream(filePath, 'utf8');
+
+    const optimized = new Map();
+    let isFirst = true;
+    let last = '';
+
+    streamReading.on('data', (chunk) => {
+      const goods = chunk.split('\n');
+
+      if (isFirst) {
+        goods.shift();
+        isFirst = false;
+      }
+      goods.unshift(...(last + goods.shift()).split('\n'));
+      last = goods.pop();
+
+      for (let index = 0; index < goods.length; index++) {
+        stringToObject(goods[index], optimized);
+      }
+    });
+
+    streamReading.on('end', () => {
+      const { result, totalQuantity } = arrayObjectsToString(optimized);
+      filePath = path.resolve('./upload/optimized/', fileName);
+
+      fs.writeFile(filePath, result, (err) => {
+        if (err) {
+          console.error('Failed to write file!', err);
+          return reject(err);
+        }
+        return resolve(totalQuantity);
+      });
+    });
+
+    streamReading.on('error', (err) => {
+      console.error('Failed to read file!', err);
+      return reject(err);
+    });
+  });
+}
+
 module.exports = {
   getDiscountCallback,
   getDiscountPromise,
   getDiscountAsyncAwait,
   createCsvToJson,
+  csvOptimization,
 };
